Extracted ListTileTheme to its own file
