+++ packages/flutter/lib/material.dart
@@ -84,6 +84,7 @@ export 'src/material/ink_well.dart';
 export 'src/material/input_border.dart';
 export 'src/material/input_decorator.dart';
 export 'src/material/list_tile.dart';
+export 'src/material/list_tile_theme.dart';
 export 'src/material/material.dart';
 export 'src/material/material_button.dart';
 export 'src/material/material_localizations.dart';
