+++ packages/flutter/lib/src/material/radio_list_tile.dart
@@ -5,6 +5,7 @@
 import 'package:flutter/widgets.dart';
 
 import 'list_tile.dart';
+import 'list_tile_theme.dart';
 import 'radio.dart';
 import 'theme.dart';
 import 'theme_data.dart';
